Using index and named imports
