+++ app/javascript/src/LoginPage/index.js
@@ -1,1 +1,4 @@
-export {SimpleLoginPage, LoginPageWrapper} from 'LoginPage/LoginPageWrapper.jsx'
+export {ForgotCredentials} from 'LoginPage/loginForms/ForgotCredentials'
+export {Login3scaleForm} from 'LoginPage/loginForms/Login3scaleForm'
+export {RequestPasswordForm} from 'LoginPage/loginForms/RequestPassWordForm.jsx'
+export {SimpleLoginPage, LoginPageWrapper} from 'LoginPage/LoginPageWrapper'
